refactor(ui): Refactor dialog, file list, help menu, panel, preview, status bar, and toast components
- Added DialogComponent for rendering confirmation, input, and message dialogs.
- Introduced FileListComponent to display file entries with proper formatting and highlighting.
- Created HelpMenuComponent to show keyboard shortcuts and their descriptions.
- Implemented PanelComponent for organizing UI panels with parent, current, and preview content.
- Developed PreviewComponent for rendering file previews with line limits and loading states.
- Added StatusBarComponent to display path, search status, and help text in the status bar.
- Implemented ToastComponent for displaying temporary messages with severity indicators.

refactor(ui): Enhance Theme integration and configuration management

- Updated Theme class to manage icon configurations and color themes more effectively.
- Refactored various components to utilize the new Theme structure for consistent styling.
- Improved configuration handling in tests to ensure independent snapshots and stability.

+++ include/expp/ui/key_handler.hpp
@@ -185,7 +185,7 @@ struct ActionContext {
 };
 
 using ActionHandler = std::function<void(const ActionContext&)>;
-using CommandId = std::uint32_t;
+using CommandId = std::uint8_t;
 constexpr CommandId kInvalidCommandId = std::numeric_limits<CommandId>::max();
 
 /**
